Development

+++ zephyr_results_publisher/publisher.py
@@ -1,21 +1,18 @@
 import json
-import logging
 import requests
-import os
 from zephyr_results_publisher.behave_to_cucumber_converter import convert_report, validate_json
 from zephyr_results_publisher.file_util import zip_file, get_path_dir
 from zephyr_results_publisher.helper import check_response_status, find_folder_id_by_name
 
-logging.basicConfig(level=logging.INFO, format='%(asctime)s :: %(levelname)s :: %(message)s')
-
 BASE_URL = "https://api.zephyrscale.smartbear.com/v2"
-ZEPHYR_TOKEN = str(os.environ.get("ZEPHYR_TOKEN"))
 
 
-def publish(project_key, source_report_file, report_format, auto_create_test_cases="true"):
+def publish(zephyr_token, project_key, source_report_file, report_format, auto_create_test_cases="true"):
+    if does_zephyr_token_exist(zephyr_token) is False:
+        return
     if report_format == "behave":
         report_format = "cucumber"
-        logging.info(f"Start converting: {source_report_file}")
+        print(f"Start converting: {source_report_file}")
         report = convert_behave_report(source_report_file)
         if report is not None:
             with open(source_report_file, "w") as f:
@@ -33,21 +30,22 @@ def publish(project_key, source_report_file, report_format, auto_create_test_cas
         "autoCreateTestCases": auto_create_test_cases
     }
     headers = {
-        "Authorization": "Bearer " + ZEPHYR_TOKEN
+        "Authorization": "Bearer " + zephyr_token
     }
     files = {
         "file": open(output_zip, 'rb')
     }
 
-    logging.info(f"Sending results to Zephyr Scale...")
+    print(f"Sending results to Zephyr Scale...")
     response = requests.post(url, files=files, params=params, headers=headers)
     check_response_status(response, 200)
     parsed_response = json.loads(response.text)
-    logging.info(f"Parsed response: {parsed_response}")
+    print(f"Parsed response: {parsed_response}")
     return parsed_response
 
 
-def publish_customized_test_cycle(project_key,
+def publish_customized_test_cycle(zephyr_token,
+                                  project_key,
                                   source_report_file,
                                   report_format,
                                   auto_create_test_cases="true",
@@ -56,7 +54,11 @@ def publish_customized_test_cycle(project_key,
                                   test_cycle_description="",
                                   test_cycle_jira_project_version=1,
                                   test_cycle_custom_fields=None):
-    if test_cycle_custom_fields is None:
+    if does_zephyr_token_exist(zephyr_token) is False:
+        return
+    if test_cycle_custom_fields is None \
+            or test_cycle_custom_fields == "" \
+            or test_cycle_custom_fields == "{}":
         test_cycle_custom_fields = {}
 
     if report_format == "behave":
@@ -73,7 +75,8 @@ def publish_customized_test_cycle(project_key,
     output_zip = f"{source_path_dir}/testResults.zip"
     zip_file(source_report_file, output_zip)
 
-    test_cycle = customize_test_cycle(project_key,
+    test_cycle = customize_test_cycle(zephyr_token,
+                                      project_key,
                                       test_cycle_name,
                                       test_cycle_folder_name,
                                       test_cycle_description,
@@ -84,19 +87,26 @@ def publish_customized_test_cycle(project_key,
         "autoCreateTestCases": auto_create_test_cases
     }
     headers = {
-        "Authorization": "Bearer " + ZEPHYR_TOKEN
+        "Authorization": "Bearer " + zephyr_token
     }
     files = {
         "file": open(output_zip, 'rb'),
         "testCycle": ("test_cycle.json", test_cycle, "application/json")
     }
 
-    logging.info(f"Sending results to Zephyr Scale...")
+    print(f"Sending results to Zephyr Scale...")
     response = requests.post(url, files=files, params=params, headers=headers)
     check_response_status(response, 200)
     parsed_response = json.loads(response.text)
-    logging.info(f"Parsed response: {parsed_response}")
+    print(f"Parsed response: {parsed_response}")
     return parsed_response
+
+
+def does_zephyr_token_exist(zephyr_token):
+    if zephyr_token is None:
+        print("Zephyr Scale API Access Token is not provided. Results will not be published to Zephyr Scale.")
+        return False
+    return True
 
 
 def convert_behave_report(report_path):
@@ -111,11 +121,16 @@ def validate_report_schema(report_path):
     validate_json(report)
 
 
-def customize_test_cycle(project_key, test_cycle_name="Automation cycle", folder_name="All test cycles",
-                         description="", jira_project_version=1, custom_fields=None):
+def customize_test_cycle(zephyr_token,
+                         project_key,
+                         test_cycle_name="Automation cycle",
+                         folder_name="All test cycles",
+                         description="",
+                         jira_project_version=1,
+                         custom_fields=None):
     if custom_fields is None:
         custom_fields = {}
-    folder_id = get_folder_id_by_name(folder_name, project_key, 20)
+    folder_id = get_folder_id_by_name(zephyr_token, folder_name, project_key, 20)
     test_cycle_json = {
         "name": test_cycle_name,
         "description": description,
@@ -123,11 +138,11 @@ def customize_test_cycle(project_key, test_cycle_name="Automation cycle", folder
         "folderId": folder_id,
         "customFields": custom_fields
     }
-    logging.info(f"Custom test cycle is generated: {test_cycle_json}")
+    print(f"Custom test cycle is generated: {test_cycle_json}")
     return json.dumps(test_cycle_json)
 
 
-def get_folder_id_by_name(name, project_key, max_results):
+def get_folder_id_by_name(zephyr_token, name, project_key, max_results):
     url = BASE_URL + f"/folders"
 
     params = {
@@ -137,10 +152,8 @@ def get_folder_id_by_name(name, project_key, max_results):
         "maxResults": max_results
     }
     headers = {
-        "Authorization": "Bearer " + ZEPHYR_TOKEN
+        "Authorization": "Bearer " + zephyr_token
     }
-
-    logging.info(params)
 
     response = requests.get(url, params=params, headers=headers)
     check_response_status(response, 200)
